bgpd: fix crashes with conditional advertisement on vpn/evpn safi
The "advertise-map" feature causes crashes when configured on vpn or
evpn. For example:

> router bgp 65500
>  address-family ipv4 vpn
>  neighbor 10.125.0.2 activate
>  neighbor 10.125.0.2 advertise-map RMAP-COND exist-map RMAP-EXIST

L3VPN or EVPN BGP table has two levels of table. The current code tries
to access to bgp_node on the first level of table, which causes crashes.

When SAFI is vpn or evpn, parse the second level of table.

Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>
Signed-off-by: Louis Scalbert <louis.scalbert@6wind.com>