bgpd, lib: add route-target constraint prefix-list
Add an Route-Target Constraint (RTC) prefix-list type. It will be used
in the next commits to compare the route-target attributes of L3VPN and
EVPN prefixes against its route-target values.

The prefix_match() function is modified so that, for RTC prefix-lists,
only the route-target is compared, ignoring the origin AS value. The
origin AS is still compared by prefix_match() in the other cases. It is
used in RTC BGP tables to track which AS requested the route-target,
ensuring that only prefixes from the corresponding AS are withdrawn when
necessary.

Signed-off-by: Alexander Sohn <github@asohn.de>
Signed-off-by: Louis Scalbert <louis.scalbert@6wind.com>