staticd,yang: deprecate multi-hop bfd monitor leaf
multi-hop yang leaf is used to configure whether the BFD session should
use the single or multi hop mode. By default, it is created and set to
false (single-hop mode).

In the next commits, the auto-hop mode will be added the single or multi
hop mode to let staticd choose automatically the single or multi hop
depending on nexthop tracking. An enumeration leaf will be more
appropriate to deal with this three hop mode cases.

Deprecate the multi-hop and do not create it by default. No value means
single-hop for now.

Signed-off-by: Louis Scalbert <louis.scalbert@6wind.com>