bgpd: add rtc eor marker meta queue
Add a route-target Constraint (RTC) End of RIB (EoR) meta queue, to be
utilized in the next commit.

This queue is designed to hold pseudo bgp_dest entries that serve as
signals indicating the completion of a set of RTC updates. It operates
with lower priority, ensuring that actual bgp_dest entries are processed
first. These actual entries include RTC updates that are used to compute
peer route-target prefix-list. Once the marker is dequeued, L3VPN and
EVPN announcements to peers can be re-evaluated.

This mechanism is analogous to the End-of-Initialization (EOI) marker
meta queue, which signals the completion of the BGP update delay period,
allowing received UPDATE messages to be processed for RIB installation
and announcements.

Signed-off-by: Louis Scalbert <louis.scalbert@6wind.com>